Finished and tested pdfdev.c.  Very close to having a complete program. Only need to have dvi parset call routines in pdfdev.c now.

+++ dvipdf-old/numbers.h
@@ -2,6 +2,7 @@
 #define NUMBERS_H
 
 #include <stdio.h>
+#include <math.h>
 
 typedef int UNSIGNED_BYTE, SIGNED_BYTE, SIGNED_PAIR;
 typedef unsigned UNSIGNED_PAIR;
@@ -24,7 +25,7 @@ SIGNED_QUAD get_signed_quad (FILE *);
 
 UNSIGNED_QUAD get_unsigned_quad (FILE *);
 
-#define ROUND(n,acc) (floor((n)/(acc)+0.5)*(acc)) 
+#define ROUND(n,acc) ((double)floor((n)/(acc)+0.5)*(acc)) 
 
 #endif /* NUMBERS_H */
